+++ gprof/i386.c
@@ -55,7 +55,8 @@ i386_find_call (parent, p_lowpc, p_highpc)
       p_highpc = s_highpc;
     }
   DBG (CALLDEBUG, printf ("[findcall] %s: 0x%lx to 0x%lx\n",
-			  parent->name, p_lowpc, p_highpc));
+			  parent->name, (unsigned long) p_lowpc,
+			  (unsigned long) p_highpc));
 
   delta = (bfd_vma) core_text_space - core_text_sect->vma;
 
@@ -85,7 +86,8 @@ i386_find_call (parent, p_lowpc, p_highpc)
 		   *      a hit
 		   */
 		  DBG (CALLDEBUG,
-		       printf ("\tdestpc 0x%lx (%s)\n", destpc, child->name));
+		       printf ("\tdestpc 0x%lx (%s)\n",
+			       (unsigned long) destpc, child->name));
 		  arc_add (parent, child, (unsigned long) 0);
 		  instructp += 4;	/* call is a 5 byte instruction */
 		  continue;
